pyproject.toml
[project]
name = "django2"
version = "0.1.0"
description = ""
authors = [
    {name = "209512",email = "209512mon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "ipython (>=9.4.0,<10.0.0)",
    "django-summernote (>=0.8.20.0,<0.9.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
